+++ lib/terminal/terminal_readline.go
@@ -1,6 +1,6 @@
 //go:build darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris || windows
 
-package query
+package terminal
 
 import (
 	"context"
@@ -9,6 +9,7 @@ import (
 	"path/filepath"
 
 	"github.com/mithrandie/csvq/lib/cmd"
+	"github.com/mithrandie/csvq/lib/query"
 
 	"github.com/mitchellh/go-homedir"
 	"github.com/mithrandie/readline-csvq"
@@ -20,10 +21,10 @@ type ReadLineTerminal struct {
 	prompt    *Prompt
 	env       *cmd.Environment
 	completer *Completer
-	tx        *Transaction
+	tx        *query.Transaction
 }
 
-func NewTerminal(ctx context.Context, scope *ReferenceScope) (VirtualTerminal, error) {
+func NewTerminal(ctx context.Context, scope *query.ReferenceScope) (query.VirtualTerminal, error) {
 	fd := int(scope.Tx.Session.ScreenFd())
 
 	limit := *scope.Tx.Environment.InteractiveShell.HistoryLimit