+++ docs/changelog.md
@@ -5,6 +5,24 @@ title: Change Log - csvq
 
 # Change Log
 
+## Version 1.17.5
+
+Released on Jun 26, 2022
+
+- Fix a bug of the CASE expression. ([GitHub #78](https://github.com/mithrandie/csvq/issues/78))
+
+## Version 1.17.4
+
+Released on Jun 25, 2022
+
+- Add the IF NOT EXISTS clause to the syntax of the CREATE TABLE statement. ([GitHub #77](https://github.com/mithrandie/csvq/issues/77))
+
+## Version 1.17.3
+
+Released on Jun 19, 2022
+
+- Set the default format to csv when the output is piped. ([GitHub #71](https://github.com/mithrandie/csvq/issues/71))
+
 ## Version 1.17.1
 
 Released on Jun 17, 2022